+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.0.2] 22-03-2023
 
+### Added
+- /routes/HomeStack.js - routing for screens
+- /screens/Home.js - home screen
+- /screens/Profile.js - profile screen
+- /screens/Login.js - login screen with inputs and button routing to Home.js
+
+### Modified
+- App.js
+  
+  
 ## [0.0.1] 22-03-2023
 
 ### Added
-- CHANGELOG.MD
-- README.MD
+- CHANGELOG.md
+- README.md